+++ src/org/ojalgo/type/keyvalue/StringToObject.java
@@ -23,6 +23,10 @@ package org.ojalgo.type.keyvalue;
 
 import org.ojalgo.netio.ASCII;
 
+/**
+ * @deprecated v49 Use {@link EntryPair.ObjectObject} instead.
+ */
+@Deprecated
 public final class StringToObject<V> implements KeyValue<String, V> {
 
     public final String key;